pyproject.toml
[project]
name = "dadam-saas"
version = "0.1.0"
description = "다담 SaaS - AI 주문제작 가구 시뮬레이션 & 견적 플랫폼"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.34.0",
    "supabase==2.12.0",
    "stripe>=14.0.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.1",
    "Pillow>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
